fix(section, constants, popupwithimage): move constants, add object
move constants.js to utils, rename metod renrerAll to renderItems (section.js), add this._newPlacePopup for search of the element (PopupWithImage.js)
