+++ src/components/Section.js
@@ -10,7 +10,7 @@ export default class Section {
     this._container.prepend(this._renderer(item));
   }
 
-  renderAll() {
+  renderItems() {
     this._items.forEach(item => {
       this.addItem(item);
     });
